gpt_misc: Return uint64_t from GptGetEntrySize functions Android Bringup: See http://go/android-fw-sync
Change GptGetEntrySizeBytes and GptGetEntrySizeLba return type to
uint64_t to prevent overflows for partitions with large size.

BUG=None
TEST=Print partition size using fastboot getvar all
BRANCH=main

(cherry picked from commit 1f7ca823da09f6c8cc451091a532b0cd60d90c34)

Original-Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5922816
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
GitOrigin-RevId: 1f7ca823da09f6c8cc451091a532b0cd60d90c34
Cr-Build-Id: 8731843853558386257
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731843853558386257
Copybot-Job-Name: android-main-vboot_reference-copybot-downstream
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6003746
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>

+++ firmware/lib/gpt_misc.c
@@ -233,7 +233,7 @@ int IsUnusedEntry(const GptEntry *e)
  * Desc: This function returns size(in lba) of a partition represented by
  * given GPT entry.
  */
-size_t GptGetEntrySizeLba(const GptEntry *e)
+uint64_t GptGetEntrySizeLba(const GptEntry *e)
 {
 	return (e->ending_lba - e->starting_lba + 1);
 }
@@ -243,7 +243,7 @@ size_t GptGetEntrySizeLba(const GptEntry *e)
  * Desc: This function returns size(in bytes) of a partition represented by
  * given GPT entry.
  */
-size_t GptGetEntrySizeBytes(const GptData *gpt, const GptEntry *e)
+uint64_t GptGetEntrySizeBytes(const GptData *gpt, const GptEntry *e)
 {
 	return GptGetEntrySizeLba(e) * gpt->sector_bytes;
 }
